fix schedule listening & adhan background notification
