+++ android/app/src/main/AndroidManifest.xml
@@ -36,19 +36,6 @@
           android:foregroundServiceType="mediaPlayback"
           >
     </service>
-    <service android:name="com.ryanheise.audioservice.AudioService"
-     android:exported="true">
-      <intent-filter>
-        <action android:name="android.media.browse.MediaBrowserService" />
-      </intent-filter>
-    </service>
-
-    <receiver android:name="com.ryanheise.audioservice.MediaButtonReceiver" 
-     android:exported="true">
-      <intent-filter>
-        <action android:name="android.intent.action.MEDIA_BUTTON" />
-      </intent-filter>
-    </receiver> 
     <activity
       android:name=".MainActivity"
       android:configChanges="orientation|keyboardHidden|keyboard|screenSize|smallestScreenSize|locale|layoutDirection|fontScale|screenLayout|density|uiMode"
