+++ lib/main.dart
@@ -28,7 +28,7 @@ import 'package:mawaqit/src/helpers/riverpod_sentry_provider_observer.dart';
 import 'package:mawaqit/src/pages/SplashScreen.dart';
 import 'package:mawaqit/src/services/audio_manager.dart';
 import 'package:mawaqit/src/services/FeatureManager.dart';
-import 'package:mawaqit/src/services/notification_background_service.dart';
+import 'package:mawaqit/src/services/background_services.dart';
 import 'package:mawaqit/src/services/mosque_manager.dart';
 import 'package:mawaqit/src/services/theme_manager.dart';
 import 'package:mawaqit/src/services/toggle_screen_feature_manager.dart';
@@ -105,7 +105,7 @@ Future<void> _initializePermissions() async {
   final deviceModel = await _getDeviceModel();
 
   await _handleOverlayPermissions(deviceModel, isRooted);
-  await NotificationBackgroundService.initializeService();
+  await UnifiedBackgroundService.initializeService();
 }
 
 Future<String> _getDeviceModel() async {
@@ -136,7 +136,7 @@ class _MyAppState extends riverpod.ConsumerState<MyApp> with WidgetsBindingObser
   void initState() {
     super.initState();
     WidgetsBinding.instance.addObserver(this);
-    NotificationBackgroundService.setNotificationVisibility(false);
+    UnifiedBackgroundService.setNotificationVisibility(false);
   }
 
   @override
@@ -147,7 +147,7 @@ class _MyAppState extends riverpod.ConsumerState<MyApp> with WidgetsBindingObser
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
-    NotificationBackgroundService().didChangeAppLifecycleState(state);
+    UnifiedBackgroundService().didChangeAppLifecycleState(state);
   }
 
   @override
