+++ lib/src/helpers/AppDate.dart
@@ -30,8 +30,8 @@ class AppDateTime {
     _initialRealTime.year,
     11,
     1,
-    00,
-    -20,
+    15,
+    45,
     00,
   );
