+++ lib/src/services/mosque_manager.dart
@@ -28,7 +28,6 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import '../data/data_source/device_info_data_source.dart';
 import '../helpers/AppDate.dart';
-import 'notification_background_service.dart';
 import 'mixins/audio_mixin.dart';
 import 'mixins/connectivity_mixin.dart';
