+++ lib/src/services/notification/notification_service.dart
@@ -4,12 +4,12 @@ import 'package:notification_overlay/notification_overlay.dart';
 class NotificationService {
   static Timer? _notificationTimer;
 
-  static Future<void> showPrayerNotification(String salahName, String prayerTime) async {
+  static Future<void> showPrayerNotification(String salahName, String prayerTime, bool shouldPlayAdhan) async {
     await dismissNotification();
     await NotificationOverlay.showNotification(
       '$salahName time ($prayerTime) notification',
     );
-    _scheduleNotificationDismissal();
+    if (!shouldPlayAdhan) _scheduleNotificationDismissal();
   }
 
   static Future<void> dismissNotification() async {
@@ -23,7 +23,7 @@ class NotificationService {
   static void _scheduleNotificationDismissal() {
     _notificationTimer?.cancel();
     _notificationTimer = Timer(
-      const Duration(minutes: 2),
+      const Duration(minutes: 1),
       dismissNotification,
     );
   }
